Registro mercado: cantidad admite decimales (kg/fracciones)
La cantidad comprada pasa de entero a decimal(10,2) para registrar
fracciones (1,5 kg, 0,25, etc.). Cubre las dos pantallas que escriben
en registros_mercado: registro directo y registro desde lista-mercado.

- Migracion ALTER cantidad a DECIMAL(10,2) (raw, dbal4 rompe ->change)
- Cast decimal:2 + accessor cantidad_formateada (es-CO, sin ceros)
- Validacion numeric/decimal:0,2 en los 3 Form Requests
- Service y controller pasan cantidad como float
- Inputs con step=any, inputmode=decimal, min=0.01
- Dashboard muestra cantidad_formateada; unitario sigue correcto

+++ app/Http/Controllers/ListaMercadoItemController.php
@@ -39,7 +39,7 @@ class ListaMercadoItemController extends Controller
         try {
             $this->session->registrarItem(
                 $item,
-                (int) $request->validated('cantidad'),
+                (float) $request->validated('cantidad'),
                 (int) $request->validated('valor'),
             );
         } catch (DomainException $e) {
